Enable pip install
Previously no entry points were defined for the OVB scripts because
they were intended to be used directly out of the bin directory.
However, because of changes in the way Python 3 handles relative
imports we can't continue to do that going forward, so we need to
be able to pip install OVB and have callable binaries.

This adds entry points for the necessary user-called files, but
does not yet update the documentation as the previous method
should still work. Once the pip version has been more thoroughly
tested the docs will be updated to the new method.

Change-Id: I158f58a58314c985baaeb7309cee6f284cb53985