Fix environment-generator
When the pep8 check that environments are up to date was added it
added a mandatory output_path parameter. This wasn't intentional
as it should have still defaulted to 'environments' and it broke the
genconfig tox target.

This change correctly makes the parameter optional so both the pep8
check and genconfig will work as expected.

Change-Id: I2e80956f3c433d2f09966364a7d53cf56114ea21