Make BMC optional
When using pre-deployed servers, you may want all of the networking
setup of OVB but don't actually need to control the instances via
IPMI. While this could already be done, it left a useless BMC
instance lying around. This change allows the BMC to be disabled
completely to clean up such environments.

Change-Id: Icd6936977684d178277ebb721a7fbb3ffad51d9a