Add embedded iPXE script to retry
Currently when the iPXE boot fails, the machine halts and requires
manual intervention to retry. In some environments when machine
boots, but the network is not yet ready to receive the DHCP request
on time. Sometimes when DHCP is ready, the filename option is not
present because the back-end service is not yet ready to provide it.
Then also sometimes the option is there, but it fails.

This patch implements a script which will do ten attempts to get
DHCP and the filename option before failing. When it fails, it will
then reboot the machine. The purpose of this is to maximise the chances
of succeeding the iPXE boot.

A small change to the element deps is also made to ensure that a boot
partition is created. This is necessary due to changes in
diskimage-builder.

Finally, the git SHA is updated in the make file to the current HEAD
to ensure that the image built is current.

Change-Id: Iccef8ff21b64f20f4a185c654c0308478eecfb4a