Don't use diskimage-builder to build IPXE image
IPXE's build process already creates bootable images, so we just use one of
those. This builds much faster, has far fewer dependencies, and produces
a much smaller image. The largest difference in output is that the
resulting image no longer uses grub, but as we weren't using any
features of grub this also means a faster boot time.

For reference, a full build takes 62 seconds on my laptop. The new
Makefile also handles dependencies much better so, for example, an
incremental rebuild after changing the embedded script takes under 3
seconds.

The ipxe repo changes to being a submodule, which continues to allow us
to pin it to a specific commit.

We no longer build the -41 image, as that was oddly specific to some
undescribed use-case. It's also trivial for whoever needed it to do
themselves if the use case is still relevant.

Change-Id: I5a6a990e6966d6eeaa2aa87cd89bfbac6738f6ee