Explicitly set MTU on interfaces in IPXE image
IPXE boot frequently hangs during introspection on upshift downloading
agent.kernel and/or agent.ramdisk because our virtualised network has a
smaller MTU than the default of 1500, and the introspection dhcp server
doesn't provide it explicitly. This is a hack to work round this until a
proper fix lands for bug 1845487, and it's backported everywhere we need
it. We should eventually aim to revert it.

Related-Bug: 1845487

Change-Id: I5efeaba8f4184c4a51575a3fb29804fd744d13b9