bmc: assume powering-on state is ACTIVE
Inside newer deployments of OpenStack, it's possible that powering
on takes quite sometime especially with a large amount of ports
attached to the VM.  This is because Nova has to unplug and plug
all of the ports which can take a period of time.

During the period of time that Nova is starting the instance up,
the state of the VM is 'powering-on' and eventually it becomes
ACTIVE.  However, the IPMI status value presented is still offline
which is inaccurate.

In order to model this more towards a real bare metal environment,
we now maintain that 'powering-up' means the instance is active
which will let tools be able to move on with other steps like
PXE boot which was going to take sometime anyways.

If the booting fails, the instance will likely timeout and fail to
provision anyways, so we don't really lose much and on the next
refresh we'll figure out that state is SHUTOFF with no task state.

Change-Id: Ifa4d0152ec3086ef2e2b65e03a4499798964b78e