Use OS::Heat::None for extra role BMC
The example's environments/base-extra-node.yaml
and environments/base-extra-node-all.yaml uses a
CentOS image to create undercloud like extra nodes.

There is no need for a virtual bmc for the servers
in this role. Set the BMC to OS::Heat::None so that
no BMC instance is created for this role.

Change-Id: I502de0a2e0352078e8f187cc0592f49ec0a9d65d